+++ include/Mesh.h
@@ -9,11 +9,51 @@
 #define _MESH_H_
 
 #include <map>
+#include <vector>
+#include <glm/mat4x4.hpp>
 #include <glm/vec2.hpp>
 #include <glm/vec3.hpp>
 
-class Mesh {
+struct IndexedRectangle {
+    unsigned int v1, v2, v3, v4; // Vertex list indices
+    unsigned int texture; // Index into object-texture list
 
+    IndexedRectangle(unsigned int t, unsigned int _v1,
+                     unsigned int _v2, unsigned int _v3, unsigned int _v4 = 0)
+        : v1(_v1), v2(_v2), v3(_v3), v4(_v4), texture(t) { }
+};
+
+struct IndexedColoredRectangle {
+    unsigned int v1, v2, v3, v4;
+    unsigned char r, g, b;
+
+    IndexedColoredRectangle(unsigned char _r, unsigned char _g, unsigned char _b,
+                            unsigned int _v1, unsigned int _v2,
+                            unsigned int _v3, unsigned int _v4 = 0)
+        : v1(_v1), v2(_v2), v3(_v3), v4(_v4), r(_r), g(_g), b(_b) { }
+};
+
+// --------------------------------------
+
+class Mesh {
+  public:
+    Mesh(const std::vector<glm::vec3>& vertices,
+         const std::vector<IndexedRectangle>& rectangles,
+         const std::vector<IndexedRectangle>& triangles,
+         const std::vector<IndexedColoredRectangle>& coloredRectangles,
+         const std::vector<IndexedColoredRectangle>& coloredTriangles);
+    void prepare();
+    void display(glm::mat4 model, glm::mat4 view, glm::mat4 projection);
+
+  private:
+    std::vector<unsigned short> indices;
+    std::vector<glm::vec3> vertices;
+    std::vector<glm::vec2> uvs;
+    std::vector<unsigned int> textures;
+
+    std::vector<unsigned short> indicesColor;
+    std::vector<glm::vec3> verticesColor;
+    std::vector<glm::vec3> colors;
 };
 
 // --------------------------------------
@@ -24,7 +64,21 @@ struct PackedVertex {
     unsigned int tex;
 
     PackedVertex(glm::vec3 p, glm::vec2 u, unsigned int t) : pos(p), uv(u), tex(t) { }
-    bool operator<(const PackedVertex& v) const { return memcmp(this, &v, sizeof(PackedVertex)) > 0; }
+
+    bool operator<(const PackedVertex& v) const {
+        return memcmp(this, &v, sizeof(PackedVertex)) > 0;
+    }
+};
+
+struct PackedColoredVertex {
+    glm::vec3 pos;
+    glm::vec3 col;
+
+    PackedColoredVertex(glm::vec3 p, glm::vec3 c) : pos(p), col(c) { }
+
+    bool operator<(const PackedColoredVertex& v) const {
+        return memcmp(this, &v, sizeof(PackedColoredVertex)) > 0;
+    }
 };
 
 template <typename T>