+++ include/StaticMesh.h
@@ -8,42 +8,14 @@
 #ifndef _STATIC_MODEL_H_
 #define _STATIC_MODEL_H_
 
-#include <vector>
-
-class TexturedTriangle {
-  public:
-    TexturedTriangle(int i[3], float s[6], int tex, unsigned short trans);
-    void display(float* vertices, float* colors, float* normals);
-
-  private:
-    int index[3];
-    float st[6];
-    int texture;
-    unsigned short transparency;
-};
-
 class StaticMesh {
   public:
-    StaticMesh(int id, int mesh);
-    ~StaticMesh();
-    void display();
-    float getRadius();
+    StaticMesh(int i, int m) : id(i), mesh(m) { }
+    void display(glm::mat4 model, glm::mat4 view, glm::mat4 projection);
+
+    int getID() { return id; }
 
   private:
-    bool dontshow;
-    float center[3];
-    float radius;
-
-    unsigned int vertexCount;
-    unsigned int colorCount;
-    unsigned int normalCount;
-
-    float* vertices;
-    float* colors;
-    float* normals;
-
-    std::vector<TexturedTriangle*> triangles;
-
     int id;
     int mesh;
 };